+++ CHANGELOG.md
@@ -11,6 +11,9 @@ Changes to the project.
 - Add `bitcoin_txcount` stat for total number of transactions since the genesis block.
   - [Source 1](https://github.com/EchterAgo/bitcoin-prometheus-exporter/commit/b368138574641e4e26a8a2dfc8be6eede82f4a73)
   - [Source 2](https://github.com/EchterAgo/bitcoin-prometheus-exporter/commit/a9a6b250f463906c1b1f9446d9a689c78c4add6d)
+- Support for Bitcoin Cash nodes ([#22][pr-22]).
+
+[pr-22]: https://github.com/jvstein/bitcoin-prometheus-exporter/pull/22
 
 ### Changed
 - Default port changed from `8334` to `9332` to avoid conflicts with Bitcoin Tor port.
@@ -19,6 +22,7 @@ Changes to the project.
   - [Source 2](https://github.com/EchterAgo/bitcoin-prometheus-exporter/commit/89212072386307fcb6a9f062ee7f958a266b1075)
 - Improve performance of block statistics.
   - [Source](https://github.com/EchterAgo/bitcoin-prometheus-exporter/commit/9c018bf081bfdc604af03d8dedd125197401b2de)
+- The number of blocks for the `bitcoin_hashps` metrics is now configurable via `HASHPS_BLOCKS` ([#22][pr-22]).
 
 [issue-12]: https://github.com/jvstein/bitcoin-prometheus-exporter/issues/12